% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leafloc}
\alias{leafloc}
\title{Add Leaflet map widget for point location.}
\usage{
leafloc(loc, bottom = "Stamen.Watercolor", top = "Stamen.TonerLite",
  width = "100\%", op = 0.5, zoom = 12)
}
\arguments{
\item{loc}{character or numeric, name of available point location or specific coordinates. See details.}

\item{bottom}{character, bottom layer provider tile set.}

\item{top}{character, top layer provider tile set.}

\item{width}{character, map widget width, defaults to \code{"100\%"}.}

\item{op}{numeric, opacity for bottom and top tile sets, respectively. If length-1, will be repeated.}

\item{zoom}{integer, map zoom, defaults to 12.}
}
\value{
a Leaflet map.
}
\description{
Add a Leaflet map widget centered on a point to a static or interactive document or a Shiny app.
}
\details{
This function generates a Leaflet map centered on \code{loc} where \code{loc} is the string name of a point location from the \code{snaplocs} package.
Alternatively, the function can be used more generally by providing a length-2 vector of longitude and lattitude.
This canned map widget uses two provider tile sets, each with default opacity, \code{op = 0.5}.
}
\examples{
leafloc("Anchorage")
leafloc(c(0, 52), zoom = 6)
}
