% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{snapalettes}
\alias{snapalettes}
\title{SNAP color palettes}
\usage{
snapalettes(palette = 1, n = NULL)
}
\arguments{
\item{palette}{character or integer, name or index of color palette.}

\item{n}{number of colors returned from \code{palette}. See details.}
}
\value{
a vector of hex colors.
}
\description{
Common SNAP color palettes.
}
\details{
Obtain commonly used SNAP color palettes for styling plots.
Palettes can be specified by name or index.
By default, all colors are returned (\code{n = NULL}).
If \code{n} is less than the number of colors in \code{palette}, the first \code{n} colors are returned.
}
\examples{
snapalettes()
}
