% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapplot.R
\docType{package}
\name{snapplot}
\alias{snapplot}
\alias{snapplot-package}
\title{snapplot: flexdashboards for SNAP data sets.}
\description{
Provides a suite of functions for graphing SNAP data sets.
}
\details{
\code{snapplot} supports other SNAPverse packages, namely \code{snapstat}.
It is recommended to use \code{snapstat} directly when working with SNAP data for graphing and analysis.
\code{snapplot} is part of the SNAPverse collection of R packages.
}
